+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - Fix local files outside source-paths not being linted even if opened/changed.
   - Consider filename + lastModified as checksum for external files, avoiding analyzing it if analysis is already present.
   - graalvm: Compile static linux with --musl for better compatibility. #868
+  - java: Use `XDG_CACHE_HOME` or `.cache/clojure-lsp` instead of config files for JDK cache.
 
 - Editor
   - Introduce ALPHA move-form command. #566